# generic-training-interface-main/api/tests/swagger.yaml
swagger: "2.0"
info:
  description: "This is a Generic Training "
  version: 1.0.0
  title: Generic Training Interface
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: generictrainingInterface.swagger.io
basePath: /v2
...
schemes:
  - http
paths:
...
  /course:
    post:
      tags:
        - course
      summary: Create course
      description: This can only be done by the logged in course.
      operationId: createCourse
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created course object
          required: true
          schema:
            $ref: "#/definitions/Course"
      responses:
        default:
          description: successful operation
...
definitions:
...
  Course:
    type: object
    properties:
      courseId:
        type: integer
        format: int64
      coursename:
        type: string
      coursedescription:
        type: string
      instructorid:
        type: integer
        format: int64
    xml:
      name: Course
...